# chore: remove graphql-utils and allow for the use of swc