--- a/packages/cli/src/commands/generate-graphql.ts
+++ b/packages/cli/src/commands/generate-graphql.ts
@@ -54,17 +54,15 @@ export default class GenerateGraphql extends Command {
       cwd: isCore ? undefined : tmpDir,
     })
 
-    if (!isCore) {
-      // yarn generate:copy-back expects the DESTINATION var to be present so it can copy the files to the correct directory
-      runCommandSync({
-        cmd: `DESTINATION=${coreDir} yarn generate:copy-back`,
-        errorMessage:
-          "Failed to format generated files. 'yarn format:generated' thrown errors",
-        throws: 'warning',
-        debug,
-        cwd: isCore ? undefined : tmpDir,
-      })
-    }
+    // yarn generate:copy-back expects the DESTINATION var to be present so it can copy the files to the correct directory
+    runCommandSync({
+      cmd: `DESTINATION=${coreDir} yarn generate:copy-back`,
+      errorMessage:
+        "Failed to format generated files. 'yarn format:generated' thrown errors",
+      throws: 'warning',
+      debug,
+      cwd: isCore ? undefined : tmpDir,
+    })
 
     console.log(
       `${chalk.green(