--- a/packages/core/.babelrc
+++ b/packages/core/.babelrc
@@ -1,4 +1,0 @@
-{
-  "presets": ["next/babel"],
-  "plugins": ["@faststore/graphql-utils/babel"]
-}