--- a/packages/core/@generated/index.ts
+++ b/packages/core/@generated/index.ts
@@ -1,0 +1,1 @@
+export * from './gql'