--- a/packages/core/jest.config.js
+++ b/packages/core/jest.config.js
@@ -1,6 +1,14 @@
+const { pathsToModuleNameMapper } = require('ts-jest')
+const { compilerOptions } = require('./tsconfig')
+
 /** @type {import('ts-jest').JestConfigWithTsJest} */
 module.exports = {
   preset: 'ts-jest',
   testPathIgnorePatterns: ['/node_modules/', 'cypress/'],
-  modulePaths: ['<rootDir>/test'],
+
+  /** Support importing from src/ or @generated/ folders (TS path imports) */
+  modulePaths: [compilerOptions.baseUrl], // <-- This will be set to 'baseUrl' value
+  moduleNameMapper: pathsToModuleNameMapper(
+    compilerOptions.paths /*, { prefix: '<rootDir>/' } */
+  ),
 }