--- a/packages/core/next.config.js
+++ b/packages/core/next.config.js
@@ -7,6 +7,8 @@ const storeConfig = require('./faststore.config')
  * */
 const nextConfig = {
   /* config options here */
+  /* Replaces terser by swc for minifying. It's the default in NextJS 13 */
+  swcMinify: true,
   images: {
     domains: [`${storeConfig.api.storeId}.vtexassets.com`],
     deviceSizes: [360, 540, 768, 1280, 1440],