--- a/packages/core/src/customizations/src/fragments/ClientManyProducts.ts
+++ b/packages/core/src/customizations/src/fragments/ClientManyProducts.ts
@@ -1,6 +1,6 @@
-import { gql } from '@faststore/graphql-utils'
+import { gql } from '@generated'
 
-export const fragment = gql`
+export const fragment = gql(`
   fragment ClientManyProducts on Query {
     search(
       first: $first
@@ -16,4 +16,4 @@ export const fragment = gql`
       }
     }
   }
-`
+`)