--- a/packages/core/src/customizations/src/fragments/ClientProduct.ts
+++ b/packages/core/src/customizations/src/fragments/ClientProduct.ts
@@ -1,9 +1,9 @@
-import { gql } from '@faststore/graphql-utils'
+import { gql } from '@generated'
 
-export const fragment = gql`
+export const fragment = gql(`
   fragment ClientProduct on Query {
     product(locator: $locator) {
       id: productID
     }
   }
-`
+`)