--- a/packages/core/src/customizations/src/fragments/ClientTopSearchSuggestions.ts
+++ b/packages/core/src/customizations/src/fragments/ClientTopSearchSuggestions.ts
@@ -1,6 +1,6 @@
-import { gql } from '@faststore/graphql-utils'
+import { gql } from '@generated'
 
-export const fragment = gql`
+export const fragment = gql(`
   fragment ClientTopSearchSuggestions on Query {
     search(first: 5, term: $term, selectedFacets: $selectedFacets) {
       suggestions {
@@ -10,4 +10,4 @@ export const fragment = gql`
       }
     }
   }
-`
+`)