--- a/packages/core/src/customizations/src/fragments/ServerCollectionPage.ts
+++ b/packages/core/src/customizations/src/fragments/ServerCollectionPage.ts
@@ -1,9 +1,9 @@
-import { gql } from '@faststore/graphql-utils'
+import { gql } from '@generated'
 
-export const fragment = gql`
+export const fragment = gql(`
   fragment ServerCollectionPage on Query {
     collection(slug: $slug) {
       id
     }
   }
-`
+`)