--- a/packages/core/src/customizations/src/fragments/ServerProduct.ts
+++ b/packages/core/src/customizations/src/fragments/ServerProduct.ts
@@ -1,9 +1,9 @@
-import { gql } from '@faststore/graphql-utils'
+import { gql } from '@generated'
 
-export const fragment = gql`
+export const fragment = gql(`
   fragment ServerProduct on Query {
     product(locator: $locator) {
       id: productID
     }
   }
-`
+`)