--- a/packages/core/src/pages/[...slug].tsx
+++ b/packages/core/src/pages/[...slug].tsx
@@ -1,7 +1,7 @@
 import { isNotFoundError } from '@faststore/api'
-import { gql } from '@faststore/graphql-utils'
 import type { GetStaticPaths, GetStaticProps } from 'next'
 
+import { gql } from '@generated'
 import type {
   ServerCollectionPageQueryQuery,
   ServerCollectionPageQueryQueryVariables,
@@ -52,7 +52,7 @@ function Page({ globalSections, type, ...otherProps }: Props) {
   )
 }
 
-const query = gql`
+const query = gql(`
   query ServerCollectionPageQuery($slug: String!) {
     ...ServerCollectionPage
     collection(slug: $slug) {
@@ -75,7 +75,7 @@ const query = gql`
       }
     }
   }
-`
+`)
 
 export const getStaticProps: GetStaticProps<
   Props,
@@ -106,7 +106,7 @@ export const getStaticProps: GetStaticProps<
       ServerCollectionPageQueryQuery
     >({
       variables: { slug },
-      operationName: query,
+      operation: query,
     }),
     getPage<PLPContentType>({
       ...(previewData?.contentType === 'plp' ? previewData : null),