--- a/packages/graphql-utils/babel.js
+++ b/packages/graphql-utils/babel.js
@@ -1,1 +1,0 @@
-module.exports = require('./dist/babel')