--- a/packages/graphql-utils/codegen.js
+++ b/packages/graphql-utils/codegen.js
@@ -1,1 +1,0 @@
-module.exports = require('./dist/codegen')