@gvc 2023-12-20
Still haven't started going through the code yet, but since you mentioned that this is a breaking change: should we update the major? How should we communicate to our customers that this change has happened without affecting them? Should we create a small migration guide?