@icazevedo 2023-12-21
> should we update the major? How should we communicate to our customers that this change has happened without affecting them? Should we create a small migration guide?

Although it would be the right thing to do according to semver, since we're still in EAP, I don't think we should update the major. Yes, we should communicate. Yes, we should create a migration guide. I'll do it!

> Can we add our gql sigil that calls the function and gives a deprecation notice
Did not get what you meant by this. Can you explain it further?