@gvc 2023-12-21
`Did not get what you meant by this. Can you explain it further?`

Sorry, sigil is the wrong name for what I was trying to say. I meant [tag function](https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Template_literals#tagged_templates).

Currently our customers use the `gql` tag function from graphl-utils to generate their queries, right? Since we're deprecating the tag function in favor of the `gql` function. Can't we recreate the tag function and call the actual function from inside that with a deprecation notice? Something like this (this is heavy pseudo-code):

```javascript
// inside graphql-utils
import { gql as new_gql } from '@generated'  

function gql(string) {
  console.warn("Using the gql tag function will be deprecated in version 2.2 please use the gql function exported from the @generated folder")
  return new_gq(string)
}
```