@icazevedo 2023-12-21
Got it! We can, but we wouldn't be able to defer to the `new_gql` as you suggested. We would have to keep the graphql-utils package as well, and I don't know if we want to do this just for the sake of a depreciation notice.